requirements.txt
requests
websocket-client
pefile
yara-python
